BakedQuadExtension impl + more misc mixins/accessors

+++ src/main/java/com/cleanroommc/blackbox/pipeline/rendering/quad/BakedQuadExtension.java
@@ -10,9 +10,9 @@ public interface BakedQuadExtension {
 
     int getColour(int index);
 
-    int getU(int index);
+    float getU(int index);
 
-    int getV(int index);
+    float getV(int index);
 
     int getFlags();
